back to micrometers till the very end, coalescence, text rendering

+++ src/main/java/karlord19/cardarchitect/Position.java
@@ -1,11 +1,11 @@
 package karlord19.cardarchitect;
 
 public class Position {
-    public float x, y;
+    public int x, y;
     public String toString() {
         return "pos: (" + x + ", " + y + ")";
     }
-    public Position(float x, float y) {
+    public Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
